Restructure how symHKLs is used in paintGrid
This change isn't directly targetting one of the main bottlenecks,
but tries to structure things to use numpy arrays in a few places
lists of numpy arrays exist, so oscillAnglesOfHKLs can be called once
for all the symHKLs and other beneficial effects. The overall time
went from 186.2 to 178.2 with this change.